+++ sources/lib.ts
@@ -45,6 +45,7 @@ export interface ITardigrade {
     get name(): DictionaryKey;
     get isAlive(): boolean;
     get mergeAgent(): Nullable<Tardigrade>;
+    get props(): Dictionary;
 }
 
 export interface ITardigradeIncidentsHandler {